Add stub for refresh_functions()
The refresh_functions() method will replace the refresh_services()
method.

It serves an equivalent purpose as refresh_services(), but is based on
functions registered on a FaaS API Gateway instead of services in a
Docker swarm. This distinction is important, as there's no guarantee
that labeled services are actually registered as functions in the
gateway.